+++ datastore/couchDB/tenantServiceDatastoreCouchDB.go
@@ -2,6 +2,7 @@ package couchDB
 
 import (
 	"fmt"
+	"strings"
 
 	"github.com/accedian/adh-gather/config"
 	ds "github.com/accedian/adh-gather/datastore"
@@ -569,11 +570,11 @@ func (tsd *TenantServiceDatastoreCouchDB) DeleteTenantThresholdProfile(tenantID 
 func (tsd *TenantServiceDatastoreCouchDB) CreateMonitoredObject(monitoredObjectReq *tenmod.MonitoredObject) (*tenmod.MonitoredObject, error) {
 	logger.Log.Debugf("Creating %s: %v\n", tenmod.TenantMonitoredObjectStr, models.AsJSONString(monitoredObjectReq))
 	monitoredObjectReq.ID = ds.GenerateID(monitoredObjectReq, string(tenmod.TenantMonitoredObjectType))
-	tenantID := ds.PrependToDataID(monitoredObjectReq.TenantID, string(admmod.TenantType))
+	baseDBID := ds.PrependToDataID(monitoredObjectReq.TenantID, strings.ToLower(string(tenmod.TenantMonitoredObjectType)))
 
-	tenantDBName := createDBPathStr(tsd.server, tenantID)
+	dbName := createDBPathStr(tsd.server, baseDBID)
 	dataContainer := &tenmod.MonitoredObject{}
-	if err := createDataInCouch(tenantDBName, monitoredObjectReq, dataContainer, string(tenmod.TenantMonitoredObjectType), tenmod.TenantMonitoredObjectStr); err != nil {
+	if err := createDataInCouch(dbName, monitoredObjectReq, dataContainer, string(tenmod.TenantMonitoredObjectType), tenmod.TenantMonitoredObjectStr); err != nil {
 		return nil, err
 	}
 	logger.Log.Debugf("Created %s: %v\n", tenmod.TenantMonitoredObjectStr, models.AsJSONString(dataContainer))
@@ -584,11 +585,11 @@ func (tsd *TenantServiceDatastoreCouchDB) CreateMonitoredObject(monitoredObjectR
 func (tsd *TenantServiceDatastoreCouchDB) UpdateMonitoredObject(monitoredObjectReq *tenmod.MonitoredObject) (*tenmod.MonitoredObject, error) {
 	logger.Log.Debugf("Updating %s: %v\n", tenmod.TenantMonitoredObjectStr, models.AsJSONString(monitoredObjectReq))
 	monitoredObjectReq.ID = ds.PrependToDataID(monitoredObjectReq.ID, string(tenmod.TenantMonitoredObjectType))
-	tenantID := ds.PrependToDataID(monitoredObjectReq.TenantID, string(admmod.TenantType))
+	baseDBID := ds.PrependToDataID(monitoredObjectReq.TenantID, strings.ToLower(string(tenmod.TenantMonitoredObjectType)))
 
-	tenantDBName := createDBPathStr(tsd.server, tenantID)
+	dbName := createDBPathStr(tsd.server, baseDBID)
 	dataContainer := &tenmod.MonitoredObject{}
-	if err := updateDataInCouch(tenantDBName, monitoredObjectReq, dataContainer, string(tenmod.TenantMonitoredObjectType), tenmod.TenantMonitoredObjectStr); err != nil {
+	if err := updateDataInCouch(dbName, monitoredObjectReq, dataContainer, string(tenmod.TenantMonitoredObjectType), tenmod.TenantMonitoredObjectStr); err != nil {
 		return nil, err
 	}
 	logger.Log.Debugf("Updated %s: %v\n", tenmod.TenantMonitoredObjectStr, models.AsJSONString(dataContainer))
@@ -599,11 +600,11 @@ func (tsd *TenantServiceDatastoreCouchDB) UpdateMonitoredObject(monitoredObjectR
 func (tsd *TenantServiceDatastoreCouchDB) GetMonitoredObject(tenantID string, dataID string) (*tenmod.MonitoredObject, error) {
 	logger.Log.Debugf("Fetching %s: %s\n", tenmod.TenantMonitoredObjectStr, dataID)
 	dataID = ds.PrependToDataID(dataID, string(tenmod.TenantMonitoredObjectType))
-	tenantID = ds.PrependToDataID(tenantID, string(admmod.TenantType))
+	baseDBID := ds.PrependToDataID(tenantID, strings.ToLower(string(tenmod.TenantMonitoredObjectType)))
 
-	tenantDBName := createDBPathStr(tsd.server, tenantID)
+	dbName := createDBPathStr(tsd.server, baseDBID)
 	dataContainer := tenmod.MonitoredObject{}
-	if err := getDataFromCouch(tenantDBName, dataID, &dataContainer, tenmod.TenantMonitoredObjectStr); err != nil {
+	if err := getDataFromCouch(dbName, dataID, &dataContainer, tenmod.TenantMonitoredObjectStr); err != nil {
 		return nil, err
 	}
 	logger.Log.Debugf("Retrieved %s: %v\n", tenmod.TenantMonitoredObjectStr, models.AsJSONString(dataContainer))
@@ -614,11 +615,11 @@ func (tsd *TenantServiceDatastoreCouchDB) GetMonitoredObject(tenantID string, da
 func (tsd *TenantServiceDatastoreCouchDB) DeleteMonitoredObject(tenantID string, dataID string) (*tenmod.MonitoredObject, error) {
 	logger.Log.Debugf("Deleting %s: %s\n", tenmod.TenantMonitoredObjectStr, dataID)
 	dataID = ds.PrependToDataID(dataID, string(tenmod.TenantMonitoredObjectType))
-	tenantID = ds.PrependToDataID(tenantID, string(admmod.TenantType))
+	baseDBID := ds.PrependToDataID(tenantID, strings.ToLower(string(tenmod.TenantMonitoredObjectType)))
 
-	tenantDBName := createDBPathStr(tsd.server, tenantID)
+	dbName := createDBPathStr(tsd.server, baseDBID)
 	dataContainer := tenmod.MonitoredObject{}
-	if err := deleteDataFromCouch(tenantDBName, dataID, &dataContainer, tenmod.TenantMonitoredObjectStr); err != nil {
+	if err := deleteDataFromCouch(dbName, dataID, &dataContainer, tenmod.TenantMonitoredObjectStr); err != nil {
 		return nil, err
 	}
 	logger.Log.Debugf("Deleted %s: %v\n", tenmod.TenantMonitoredObjectStr, models.AsJSONString(dataContainer))
@@ -628,11 +629,11 @@ func (tsd *TenantServiceDatastoreCouchDB) DeleteMonitoredObject(tenantID string,
 // GetAllMonitoredObjects - CouchDB implementation of GetAllMonitoredObjects
 func (tsd *TenantServiceDatastoreCouchDB) GetAllMonitoredObjects(tenantID string) ([]*tenmod.MonitoredObject, error) {
 	logger.Log.Debugf("Fetching all %s\n", tenmod.TenantMonitoredObjectStr)
-	tenantID = ds.PrependToDataID(tenantID, string(admmod.TenantType))
+	baseDBID := ds.PrependToDataID(tenantID, strings.ToLower(string(tenmod.TenantMonitoredObjectType)))
 
-	tenantDBName := createDBPathStr(tsd.server, tenantID)
+	dbName := createDBPathStr(tsd.server, baseDBID)
 	res := make([]*tenmod.MonitoredObject, 0)
-	if err := getAllOfTypeFromCouchAndFlatten(tenantDBName, string(tenmod.TenantMonitoredObjectType), tenmod.TenantMonitoredObjectStr, &res); err != nil {
+	if err := getAllOfTypeFromCouchAndFlatten(dbName, string(tenmod.TenantMonitoredObjectType), tenmod.TenantMonitoredObjectStr, &res); err != nil {
 		return nil, err
 	}
 
@@ -643,10 +644,10 @@ func (tsd *TenantServiceDatastoreCouchDB) GetAllMonitoredObjects(tenantID string
 // GetMonitoredObjectToDomainMap - CouchDB implementation of GetMonitoredObjectToDomainMap
 func (tsd *TenantServiceDatastoreCouchDB) GetMonitoredObjectToDomainMap(moByDomReq *tenmod.MonitoredObjectCountByDomainRequest) (*tenmod.MonitoredObjectCountByDomainResponse, error) {
 	logger.Log.Debugf("Fetching %s: %v\n", tenmod.MonitoredObjectToDomainMapStr, models.AsJSONString(moByDomReq))
-	tenantID := ds.PrependToDataID(moByDomReq.TenantID, string(admmod.TenantType))
+	baseDBID := ds.PrependToDataID(moByDomReq.TenantID, strings.ToLower(string(tenmod.TenantMonitoredObjectType)))
 
-	tenantDBName := createDBPathStr(tsd.server, tenantID)
-	db, err := getDatabase(tenantDBName)
+	dbName := createDBPathStr(tsd.server, baseDBID)
+	db, err := getDatabase(dbName)
 	if err != nil {
 		return nil, err
 	}
@@ -665,7 +666,7 @@ func (tsd *TenantServiceDatastoreCouchDB) GetMonitoredObjectToDomainMap(moByDomR
 		requestBody := map[string]interface{}{}
 		requestBody["keys"] = moByDomReq.DomainSet
 
-		fetchResponse, err = fetchDesignDocumentResults(requestBody, tenantDBName, monitoredObjectsByDomainIndex)
+		fetchResponse, err = fetchDesignDocumentResults(requestBody, dbName, monitoredObjectsByDomainIndex)
 		if err != nil {
 			return nil, err
 		}
@@ -847,9 +848,9 @@ func (tsd *TenantServiceDatastoreCouchDB) GetAllTenantThresholdProfile(tenantID 
 // BulkInsertMonitoredObjects - CouchDB implementation of BulkInsertMonitoredObjects
 func (tsd *TenantServiceDatastoreCouchDB) BulkInsertMonitoredObjects(tenantID string, value []*tenmod.MonitoredObject) ([]*common.BulkOperationResult, error) {
 	logger.Log.Debugf("Bulk creating %s: %v\n", tenmod.TenantMonitoredObjectStr, models.AsJSONString(value))
-	tenantID = ds.PrependToDataID(tenantID, string(admmod.TenantType))
-	tenantDBName := createDBPathStr(tsd.server, tenantID)
-	resource, err := couchdb.NewResource(tenantDBName, nil)
+	baseDBID := ds.PrependToDataID(tenantID, strings.ToLower(string(tenmod.TenantMonitoredObjectType)))
+	dbName := createDBPathStr(tsd.server, baseDBID)
+	resource, err := couchdb.NewResource(dbName, nil)
 	if err != nil {
 		return nil, err
 	}